+++ Sources/KeyboardKit/Images/UIImage+Photos.swift
@@ -14,6 +14,7 @@ public extension UIImage {
      Saves an image to the user's photo album. This requires
      the correct permission to be added to the `Info.plist`.
      */
+    @available(*, deprecated, message: "This function will be removed in 4.0. Copy it to your own project to keep using it.")
     func saveToPhotos(completionTarget: AnyObject?, completionSelector: Selector?) {
         UIImageWriteToSavedPhotosAlbum(self, completionTarget, completionSelector, nil)
     }
@@ -25,6 +26,7 @@ public extension UIImage {
      This function uses `StandardPhotosImageService.default`
      to save the image to photos.
      */
+    @available(*, deprecated, message: "This function will be removed in 4.0. Copy it to your own project to keep using it.")
     func saveToPhotos(completion: @escaping (Error?) -> Void) {
         let service = StandardPhotosImageService.default
         service.saveImageToPhotos(self, completion: completion)